+++ src/main/java/com/spring/timebook/auth/oauth/OAuthKakaoService.java
@@ -1,10 +1,9 @@
-package com.spring.timebook.auth;
-
+package com.spring.timebook.auth.oauth;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.spring.timebook.auth.exception.ParseException;
-import com.spring.timebook.config.OAuthNaverConfig;
+import com.spring.timebook.config.OAuthKakaoConfig;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.http.HttpEntity;
@@ -14,60 +13,52 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
-
-import java.math.BigInteger;
-import java.security.SecureRandom;
 import java.util.Map;
 
 @Getter
 @Setter
-public class OAuthNaverLoginService implements OAuthLoginService {
+public class OAuthKakaoService implements OAuthProviderService {
 
     private String clientId;
-
-    private String clientSecret;
-
     private String redirectUri;
 
-    public OAuthNaverLoginService(OAuthNaverConfig oAuthNaverConfig) {
-        this.clientId = oAuthNaverConfig.getClientId();
-        this.clientSecret = oAuthNaverConfig.getClientSecret();
-        this.redirectUri = oAuthNaverConfig.getRedirectUri();
+    public OAuthKakaoService(OAuthKakaoConfig oAuthKakaoConfig){
+        this.clientId = oAuthKakaoConfig.getApiKey();
+        this.redirectUri = oAuthKakaoConfig.getRedirectUri();
     }
 
     @Override
     public String redirect() {
-       return getRedirectUrl();
+        return getRedirectUrl();
     }
 
     @Override
     public OAuthUser process(Map<String, String> info) {
-        
-        // access token 발급
 
         String code = info.get("code");
-        String state = info.get("state");
-        String accessToken = getAccessToken(code, state);
+        String accessToken = getAccessToken(code);
 
-        // 유저 프로필 정보 가져오기
         return getUserInfo(accessToken);
     }
 
-    public String getAccessToken(String code, String state) {
+    public void logout(){
+
+    }
+
+    private String getAccessToken(String code){
         HttpHeaders headers = new HttpHeaders();
         headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
 
         MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
         body.add("grant_type", "authorization_code");
         body.add("client_id", clientId);
-        body.add("client_secret", clientSecret);
+        body.add("redirect_uri", redirectUri);
         body.add("code", code);
-        body.add("state", state);
 
         HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(body, headers);
         RestTemplate rt = new RestTemplate();
         // 토급 요청
-        String TOKEN_URI = "https://nid.naver.com/oauth2.0/token";
+        String TOKEN_URI = "https://kauth.kakao.com/oauth/token";
         ResponseEntity<String> response = rt.exchange(
                 TOKEN_URI,
                 HttpMethod.POST,
@@ -77,32 +68,25 @@ public class OAuthNaverLoginService implements OAuthLoginService {
         return parseToken(response.getBody());
     }
 
-    public OAuthUser getUserInfo(String accessToken) {
+    private OAuthUser getUserInfo(String accessToken){
         HttpHeaders headers = new HttpHeaders();
         headers.add("Authorization", "Bearer " + accessToken);
-        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+        headers.add("Content-type", "application/x-www-form-urlencoded");
 
         // HTTP 요청 보내기 - Post 방식
         HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(headers);
         RestTemplate rt = new RestTemplate();
         // 프로필 정보 요청
-        String USER_INFO_URI = "https://openapi.naver.com/v1/nid/me";
+        String userInfoUri = "https://kapi.kakao.com/v2/user/me";
         ResponseEntity<String> response = rt.exchange(
-                USER_INFO_URI,
-                HttpMethod.GET,
+                userInfoUri,
+                HttpMethod.POST,
                 request,
                 String.class
         );
 
-         return parseUserInfo(response.getBody());
+        return parseUserInfo(response.getBody());
     }
-
-    private String getRedirectUrl(){
-        String state = new BigInteger(130, new SecureRandom()).toString();
-        return "https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=" + clientId
-                + "&state=" + state + "&redirect_uri=" + redirectUri;
-    }
-
 
     private String parseToken(String responseBody){
         try{
@@ -115,29 +99,34 @@ public class OAuthNaverLoginService implements OAuthLoginService {
         }
     }
 
-    private OAuthUser parseUserInfo(String responseBody) {
+    private OAuthUser parseUserInfo(String responseBody){
         try {
             ObjectMapper objectMapper = new ObjectMapper();
-            JsonNode jsonNode = objectMapper.readTree(responseBody);
-
-            String resultCode = jsonNode.get("resultcode").asText();
-            String message = jsonNode.get("message").asText();
-
-            JsonNode response = jsonNode.get("response");
+            JsonNode response = objectMapper.readTree(responseBody);
 
             String id = response.get("id").asText();
-            String email = response.get("email").asText();
-            String nickname = response.get("name").asText();
+
+            JsonNode account = response.get("kakao_account");
+            String email = account.get("email").asText();
+
+            JsonNode profile = account.get("profile");
+            String nickname = profile.get("nickname").asText();
+
 
             return OAuthUser.builder()
                     .email(email)
                     .nickname(nickname)
-                    .snsType(OAuthProvider.NAVER)
                     .snsId(id)
+                    .snsType(OAuthProvider.KAKAO)
                     .build();
 
         } catch (Exception e) {
             throw new ParseException(e.getMessage());
         }
+    }
+
+    private String getRedirectUrl(){
+        return "https://kauth.kakao.com/oauth/authorize?response_type=code&client_id="+ clientId +
+                "&redirect_uri=" + redirectUri;
     }
 }
