+++ src/main/java/com/spring/timebook/auth/oauth/OAuthProvider.java
@@ -1,4 +1,4 @@
-package com.spring.timebook.auth;
+package com.spring.timebook.auth.oauth;
 
 public enum OAuthProvider {
     GOOGLE, NAVER, KAKAO
